+++ source/MaterialXTest/Node.cpp
@@ -12,6 +12,24 @@
 
 namespace mx = MaterialX;
 
+bool isTopologicalOrder(const std::vector<mx::ElementPtr>& elems)
+{
+    std::set<mx::ElementPtr> prevElems;
+    for (mx::ElementPtr elem : elems)
+    {
+        for (size_t i = 0; i < elem->getUpstreamEdgeCount(); i++)
+        {
+            mx::ElementPtr upstreamElem = elem->getUpstreamElement(nullptr, i);
+            if (upstreamElem && !prevElems.count(upstreamElem))
+            {
+                return false;
+            }
+        }
+        prevElems.insert(elem);
+    }
+    return true;
+}
+
 TEST_CASE("Node", "[node]")
 {
     // Create a document.
@@ -44,10 +62,6 @@ TEST_CASE("Node", "[node]")
     output2->setConnectedNode(image);
     REQUIRE(output1->getUpstreamElement() == constant);
     REQUIRE(output2->getUpstreamElement() == image);
-
-    // Set file prefix.
-    nodeGraph->setFilePrefix("prefix");
-    REQUIRE(image->getActiveFilePrefix() == "prefix");
 
     // Define and reference a custom type.
     doc->addTypeDef("spectrum");
@@ -164,29 +178,10 @@ TEST_CASE("Topological sort", "[nodegraph]")
     // Validate the document.
     REQUIRE(doc->validate());
 
-    // Blessed solution to the topological sort
-    const std::vector<std::string> blessedElemOrder =
-    {
-        "output",
-        "mix",
-        "multiply",
-        "add3",
-        "noise3d",
-        "image1",
-        "add1",
-        "add2",
-        "constant1",
-        "constant2",
-        "image2"
-    };
-
-    // Create a topological order and test it agains the blessed solution
+    // Create a topological order and validate the results.
     std::vector<mx::ElementPtr> elemOrder = nodeGraph->topologicalSort();
-    REQUIRE(elemOrder.size() == blessedElemOrder.size());
-    for (size_t i = 0; i < elemOrder.size(); ++i)
-    {
-        REQUIRE(elemOrder[i]->getName() == blessedElemOrder[i]);
-    }
+    REQUIRE(elemOrder.size() == nodeGraph->getChildren().size());
+    REQUIRE(isTopologicalOrder(elemOrder));
 }
 
 TEST_CASE("New nodegraph from output", "[nodegraph]")
@@ -237,19 +232,8 @@ TEST_CASE("New nodegraph from output", "[nodegraph]")
     out1->setConnectedNode(mix);
     out2->setConnectedNode(multiply2);
 
-    // Validate the document.
-    REQUIRE(doc->validate());
-
-    // Vector of the outputs to generate new graphs from.
+    // Generate a new graph from each output.
     std::vector<mx::OutputPtr> outputs = {out1, out2};
-
-    // Blessed solution to a topological sort for the outputs.
-    const std::vector<std::vector<std::string>> blessedElemOrder =
-    {
-        { "out1", "mix", "multiply1", "add3", "noise3d", "image1", "add1", "add2", "constant1", "constant2", "image2" },
-        { "out2", "multiply2", "noise3d", "constant3" }
-    };
-
     for (size_t i = 0; i < outputs.size(); ++i)
     {
         const mx::OutputPtr output = outputs[i];
@@ -293,14 +277,14 @@ TEST_CASE("New nodegraph from output", "[nodegraph]")
             processedNodes.insert(upstreamNode);
         }
 
-        // Create a topological order and test it agains the blessed solution.
+        // Create a topological order and validate the results.
         std::vector<mx::ElementPtr> elemOrder = nodeGraph2->topologicalSort();
-        REQUIRE(elemOrder.size() == blessedElemOrder[i].size());
-        for (size_t j = 0; j < elemOrder.size(); ++j)
-        {
-            REQUIRE(elemOrder[j]->getName() == blessedElemOrder[i][j]);
-        }
+        REQUIRE(elemOrder.size() == nodeGraph2->getChildren().size());
+        REQUIRE(isTopologicalOrder(elemOrder));
     }
+
+    // Validate the document.
+    REQUIRE(doc->validate());
 }
 
 TEST_CASE("Prune nodes", "[nodegraph]")
@@ -344,9 +328,6 @@ TEST_CASE("Prune nodes", "[nodegraph]")
     mix->setConnectedNode("bg", add3);
     mix->setConnectedNode("mask", noise3d);
     output->setConnectedNode(mix);
-
-    // Validate the document.
-    REQUIRE(doc->validate());
 
     // Set the node names we want to prune from the graph
     // and which corresponding input to use for the bypass.
@@ -399,7 +380,7 @@ TEST_CASE("Prune nodes", "[nodegraph]")
             upstreamNode = upstreamNode->getConnectedNode(inputName);
         }
 
-        if(upstreamNode)
+        if (upstreamNode)
         {
             // Get (or create) the node in the new graph.
             mx::NodePtr upstreamNode2 = nodeGraph2->getNode(upstreamNode->getName());
@@ -429,23 +410,11 @@ TEST_CASE("Prune nodes", "[nodegraph]")
         processedNodes.insert(upstreamNode);
     }
 
-    // Blessed solution to a topological sort of the
-    // prunned graph.
-    const std::vector<std::string> blessedElemOrder = 
-    {
-        "output",
-        "mix",
-        "multiply",
-        "noise3d",
-        "image1",
-        "constant1"
-    };
+    // Validate the document.
+    REQUIRE(doc->validate());
 
-    // Create a topological order and test it agains the blessed solution.
+    // Create a topological order and validate the results.
     std::vector<mx::ElementPtr> elemOrder = nodeGraph2->topologicalSort();
-    REQUIRE(elemOrder.size() == blessedElemOrder.size());
-    for (size_t i = 0; i < elemOrder.size(); ++i)
-    {
-        REQUIRE(elemOrder[i]->getName() == blessedElemOrder[i]);
-    }
+    REQUIRE(elemOrder.size() == nodeGraph2->getChildren().size());
+    REQUIRE(isTopologicalOrder(elemOrder));
 }